refactor blueprints (4/6): extract ani blueprint
Seven /ani/* routes plus all 24 ani_* helper functions move to
blueprints/ani.py — by far the biggest blueprint (~795 lines, vs.
~150 for the others).

Routes moved:
- GET /ani/weather (Cockpit HUD weather display, calls ani_get_weather)
- POST /ani/chat
- GET /ani/history
- POST /ani/clear
- POST /ani/refresh
- POST /ani/location
- GET /ani/ping

Internal helpers moved (24 functions): conversation persistence
(load/save), visit logging, ache-level tracking, session-tone
assessment, cleanup-phrase detection, recent-status/git-log/now-page
context fetchers, system-prompt + briefing builders, mood assessment,
new-day/active-hours/should-initiate logic, opener generation,
notify-publish hook, and the Grok client.

ANI_CONVERSATION_FILE and ANI_MEMORY_FILE path constants redefined
locally in blueprints/ani.py.

Cross-blueprint dependency: app.py's still-inline publish_status route
(will move to cockpit blueprint in sub-phase 5) calls ani_notify_publish.
Imported back into app.py via `from blueprints.ani import ani_notify_publish`.

Smoke-tested: /ani/weather, /ani/history, /ani/ping all 200; /publish
and /below-deck cross-blueprint 200.
